docs(layout): doxygen-style API comments, rename rebuild_indices, split validate_layout header
